Fixed problem in profiling
I cleaned the code in data generation function deleting objects I wasn't using.  I did the same in profiling script.
I added a new object with a minimum value, in this way I expect to avoid assigning 1000 to the lower boundary when it reached this value in the previous iteration